nginx: Force re-resolving of DNS
A user reported that sometimes nginx would fail to start up because it
could not resolve authrpxoy and outproxy dns records. After
investigating the issue, we were unable to reproduce it.

Nonetheless, this commit updates the NGINX config to force it to
re-resolve the DNS records at query time, instead of at container start
time. This most likely fixes the issue.